move authentication function to util
